🔧 Rename(#51): move to hooks-api directory
